+++ week8/hyunjin-crud-api/src/main/java/practice/hyunjincrudapi/comment/dto/request/UpdateCommentRequest.java
@@ -1,4 +1,4 @@
-package practice.hyunjincrudapi.comment.controller.dto.request;
+package practice.hyunjincrudapi.comment.dto.request;
 
 import jakarta.validation.constraints.NotBlank;
 import lombok.Getter;
