+++ week8/hyunjin-crud-api/src/main/java/practice/hyunjincrudapi/comment/service/CommentService.java
@@ -3,9 +3,9 @@ package practice.hyunjincrudapi.comment.service;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
-import practice.hyunjincrudapi.comment.controller.dto.request.CreateCommentRequest;
-import practice.hyunjincrudapi.comment.controller.dto.request.UpdateCommentRequest;
-import practice.hyunjincrudapi.comment.controller.dto.response.CommentResponse;
+import practice.hyunjincrudapi.comment.dto.request.CreateCommentRequest;
+import practice.hyunjincrudapi.comment.dto.request.UpdateCommentRequest;
+import practice.hyunjincrudapi.comment.dto.response.CommentResponse;
 import practice.hyunjincrudapi.comment.entity.Comment;
 import practice.hyunjincrudapi.comment.repository.JpaCommentRepository;
 import practice.hyunjincrudapi.member.entity.Member;
