+++ week8/hyunjin-crud-api/src/main/java/practice/hyunjincrudapi/member/dto/request/UpdateMemberRequest.java
@@ -1,4 +1,4 @@
-package practice.hyunjincrudapi.member.controller.dto.request;
+package practice.hyunjincrudapi.member.dto.request;
 
 import lombok.Getter;
