+++ week8/hyunjin-crud-api/src/main/java/practice/hyunjincrudapi/member/dto/response/MemberResponse.java
@@ -1,4 +1,4 @@
-package practice.hyunjincrudapi.member.controller.dto.response;
+package practice.hyunjincrudapi.member.dto.response;
 
 import lombok.AllArgsConstructor;
 import lombok.Getter;
